DRY most of the DiscoveryServerCallbacks implementations in integration tests
