+++ server/src/test/java/io/envoyproxy/controlplane/server/V2DiscoveryServerV3ResourcesAdsIT.java
@@ -1,6 +1,6 @@
 package io.envoyproxy.controlplane.server;
 
-import static io.envoyproxy.controlplane.server.V3TestSnapshots.createSnapshot;
+import static io.envoyproxy.controlplane.server.util.V3TestSnapshots.createSnapshot;
 import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
@@ -8,8 +8,6 @@ import static org.hamcrest.Matchers.containsString;
 
 import io.envoyproxy.controlplane.cache.NodeGroup;
 import io.envoyproxy.controlplane.cache.v3.SimpleCache;
-import io.envoyproxy.envoy.api.v2.DiscoveryRequest;
-import io.envoyproxy.envoy.api.v2.DiscoveryResponse;
 import io.envoyproxy.envoy.api.v2.core.Node;
 import io.grpc.netty.NettyServerBuilder;
 import io.restassured.http.ContentType;
@@ -49,35 +47,9 @@ public class V2DiscoveryServerV3ResourcesAdsIT {
         }
       });
 
-      final DiscoveryServerCallbacks callbacks = new DiscoveryServerCallbacks() {
-        @Override
-        public void onStreamOpen(long streamId, String typeUrl) {
-          onStreamOpenLatch.countDown();
-        }
-
-        @Override
-        public void onV2StreamRequest(long streamId, DiscoveryRequest request) {
-          onStreamRequestLatch.countDown();
-        }
-
-        @Override
-        public void onV3StreamRequest(long streamId,
-            io.envoyproxy.envoy.service.discovery.v3.DiscoveryRequest request) {
-          throw new IllegalStateException("Unexpected v3 request in v2 test");
-        }
-
-        @Override
-        public void onStreamResponse(long streamId, DiscoveryRequest request, DiscoveryResponse response) {
-          onStreamResponseLatch.countDown();
-        }
-
-        @Override
-        public void onV3StreamResponse(long streamId,
-            io.envoyproxy.envoy.service.discovery.v3.DiscoveryRequest request,
-            io.envoyproxy.envoy.service.discovery.v3.DiscoveryResponse response) {
-          throw new IllegalStateException("Unexpected v3 response in v2 test");
-        }
-      };
+      final DiscoveryServerCallbacks callbacks =
+          new V2OnlyDiscoveryServerCallbacks(onStreamOpenLatch, onStreamRequestLatch,
+              onStreamResponseLatch);
 
       cache.setSnapshot(
           GROUP,
